backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pika==1.3.2
motor==3.3.2
redis==5.0.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
boto3==1.34.64
python-dotenv==1.0.1
httpx==0.27.0
apache-flink==1.18.1
